@Bladrak 2025-08-13
Hello @brambaud 
Thanks for contributing! We had a PR to bump the version a few years ago, but didn't merge it due to performance issues (https://github.com/thumbor-community/aws/pull/157). 
I see that you had performance in mind when submitting this PR, did you run some tests with it?