@brambaud 2025-08-14
This branch has been rebased too.

> I see that you had performance in mind when submitting this PR, did you run some tests with it?

Yes, I ran some load tests because:
* I saw the previous attempt
* when testing https://github.com/thumbor/thumbor-aws I stumble upon the performance degradation
* when I did the upgrade, the breaking change in aibotocore was a hint that re-creating a client each time may be an issue

I don't have my machine with me right now to re-run them and paste the result because I'm on holiday but the load tests were something like:

```js
import http from "k6/http";

export default function () {
 const response = http.get("https://localhost/unsafe/800x600/filters:quality(80)/some/path/img.png");
}
````

Without the fix, this small tests (50 requests per second during 60s) was broken

```json
{
    "discardResponseBodies": true,
    "insecureSkipTLSVerify": true,
    "scenarios": {
      "test": {
        "executor": "ramping-arrival-rate",
        "startRate": 0,
        "timeUnit": "1s",
          "preAllocatedVUs": 50,
          "stages": [
          { "target": 50, "duration": "60s" }
        ]
      }
    }
  }
```

With the fix, it ran ok and a bigger tests was ok too:

```json
{
    "discardResponseBodies": true,
    "insecureSkipTLSVerify": true,
    "scenarios": {
      "test": {
        "executor": "ramping-arrival-rate",
        "startRate": 0,
        "timeUnit": "1s",
          "preAllocatedVUs": 500,
          "stages": [
          { "target": 25, "duration": "60s" },
          { "target": 25, "duration": "180s" },
          { "target": 50, "duration": "60s" },
          { "target": 50, "duration": "180s" },
          { "target": 100, "duration": "60s" },
          { "target": 100, "duration": "180s" },
          { "target": 200, "duration": "60s" },
          { "target": 200, "duration": "180s" },
          { "target": 400, "duration": "60s" },
          { "target": 400, "duration": "180s" }
        ]
      }
    }
  }
 ```

For the moment, it has not been tested on another machine than mine.

PS: In order to share the knowledge, I need to propose a patch to https://github.com/thumbor/thumbor-aws to see if the performance issues are only due to the same naïve implementation of re-using the session and using context manager to create a new client each time